fix(server): keep websocket actor helpers in flat server module
Inline websocket runtime actor parsing/routing helpers into src/modules/server.ts
and remove the extra src/modules/server/* implementation file to preserve the
flat core module convention in src/modules/AGENTS.md.

Behavior is preserved:
- UI ingress still routes to ui_core
- inference ingress still routes to inference_bridge
- websocket boundary parsing remains strict parse/try-catch with diagnostics
- server runtime paths do not silently drop invalid detail payloads

Validation:
- bun test src/modules/server/tests/server-module.spec.ts
- bun test src/inference/tests/inference-bridge.spec.ts
- bun --bun tsc --noEmit
